requirements.txt
aiogram==3.18.0
groq==0.23.1
fpdf2==2.8.3
python-dotenv==1.2.0
